pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-paper-scrape"
version = "0.1.0"
description = "ChemRxiv + OpenReview trial scraper (Aug 2025 window) → JSONL + stats."
requires-python = ">=3.11"
dependencies = [
  "httpx[http2]>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.1",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "typer>=0.12.4",
  "tqdm>=4.66.4",
  "loguru>=0.7.2",
  "tenacity>=8.3.0",
  "pandas>=2.2.2",
  "PyMuPDF>=1.24.9",
  "pytesseract>=0.3.10",
  "Pillow>=10.3.0",
  "python-docx>=1.1.2",
  "pylatexenc>=2.10",
  "pdfminer.six>=20231228",
  "ruamel.yaml>=0.18.6",
  "aiofiles>=24.1.0",
  "playwright>=1.45.0"
]

[project.scripts]
scrape = "scrape.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
